Struct snowflake::ProcessUniqueId

pub struct ProcessUniqueId {
    // some fields omitted
}

Process unique IDs are guaranteed to be unique within the current process, for the lifetime of the current process.

ID creation should be highly performant even on highly concurrent systems. It's MUCH faster than using random/time based IDs (but, on the other hand, only unique within a process).

While this crate can run out of process unique IDs, this is very unlikely assuming a sane threading model and will panic rather than potentially reusing unique IDs.

Limits

The unique ID's are sizeof(usize) + 64 bits wide and are generated by combining a usize global counter value with a 64bit thread local offset. This is important because each thread that calls new() at least once will reserve at least 264 IDs. So, the only way to run out of IDs in a reasonable amount of time is to run a 32bit system, spawn 232 threads, and claim one ID on each thread. You might be able to do this on a 64bit system but it would take a while... TL; DR: Don't create unique IDs from over 4 billion different threads on a 32bit system.

Methods

impl ProcessUniqueId

fn new() -> ProcessUniqueId

Create a new unique ID.

panics if there are no more unique IDs available. If this happens, go home and reevaluate your threading model!