Crate snowflake

A crate for quickly generating unique IDs with guaranteed properties.

This crate currently includes guaranteed process unique IDs but may include new ID types in the future.

Structs

ProcessUniqueId
Process unique IDs are guaranteed to be unique within the current process, for the lifetime of the current process.